LayoutTests/http/wpt/workers/promise-unhandled-rejection.any.worker-expected.txt

CONSOLE MESSAGE: line 1774: TypeError: null is not an object (evaluating 'this.message_target.removeEventListener')
CONSOLE MESSAGE: Unhandled Promise Rejection: Reject

Harness Error (FAIL), message = undefined

PASS UnhandledRejection event occurs if a rejected promise is not handled. 
